% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_edgeR.R
\name{DEG_edgeR}
\alias{DEG_edgeR}
\title{RNAseq differential analysis}
\usage{
DEG_edgeR(exprset.group, coef = 2, pval = 0.05, fdr = 0.1, logfc = 1)
}
\arguments{
\item{exprset.group}{x is a 3-object list: list(eset=exprset, group=phen, f_mark="")}

\item{coef}{default is 2, meanig the 2nd fator vs 1st factor of group.}

\item{pval}{default is 0.05}

\item{fdr}{default is 0.1}

\item{logfc}{default is 1}
}
\value{
a list: list(qlf = qlf, resdf = et.norm, deg = etSig)
qlf: mid data of edgeR, et.norm: all result of DEG, sig result of DEG.
}
\description{
The function is designed to perform differential analysis by edgeR.
}
\details{
Following data is requried: express, group.
}
\examples{
# Run: diffan <- DEG_edgeR(exprset.group) # pval=0.05, fdr=0.1, logfc=1
# exprset.group  由下列三个数据打包而成，示例：list(eset=exprset, group=phen, f_mark="")
# *1 exprset.group$eset： 矩阵数据格式(数值型，整型)
#       | row1 | row2 | row3  | row4
# gene1 |  34  |  23  |  56   |  23
# gene2 |  35  |  23  |  12   |  23
# gene3 |  12  |  78  |  78   |  78
# *2 exprset.group$group： 分组数据格式：建议只有一列数据
#    # 需要组的行名=表达谱的列名 rownames(group) == colname(eset)
#            type
# rowname1 | tumor
# rowname2 | tumor
# rowname3 | normal
# rowname4 | normal
# *3 exprset.group$f_mark：list标记项，可以为空，但是如果是批量差异分析，建议必须要有此参数，方便后续对个结果进行保存和导出
}
