% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_edgeR.R
\name{DEG_preEset.dup}
\alias{DEG_preEset.dup}
\title{removes duplicate row from a dataframe}
\usage{
DEG_preEset.dup(df, index = "gene_name")
}
\arguments{
\item{df}{dataframe the first of the dataframe is the rowindex and recommend to set name as gene_name.}

\item{index}{character default is "gene_name", if the fist column of df is not gene_name, please set the column name.}
}
\value{
dataframe a express eset which rowname is the unduplicated index
}
\description{
removes duplicate row from a dataframe, retains the row with the higher mean.
}
\examples{
# eset <- DEG_preEset.dup(df = eset) # index = "gene_name"
# df:
# |gene_name | row1 | row2 | row3  | row4
# | gene1    |  34  |  23  |  56   |  23
# | gene2    |  35  |  23  |  12   |  23
# | gene3    |  12  |  78  |  78   |  78
}
